popup: only hand off focus if we still have it.
When closing a popup we move the focus to the target - but should only
do this if we still have the focus.
It is too late to check if we have it directly as pane_close() has moved
it.  So pass a flag to "Close" to report if the pane previously had focus.

Signed-off-by: NeilBrown <neil@brown.name>