renderline: don't resize a pane after clearing it.
When we clear a pane, the display manager remembers the size.
So changing the size again - without clearing again - can cause
problems.

Fix this in two places: renderline avoids the resize and ncurses guards
against it.

There is still a display problem when the last line of a file wraps, but
at least it doesn't crash.

Signed-off-by: NeilBrown <neil@brown.name>