search: don't depend on attributes to find match
When re-highlighting after a scroll, we assume the current match will
already be highlighted, but it might not be.
If we find a match that is, leave it. If we find a match that isn't
highlight, check the start mark - which we now take a copy of
when highlighting the match.

Signed-off-by: NeilBrown <neil@brown.name>