pane: fix problem with searchbox positioning.
When the search target is found near the search box, the searchbox can
move to the size so as not to obscure it match.
Sometimes it moves the wrong way.

This was because pane_mapxy does not work when the target is not on a
direct line to the root from the focus.

In this case both the target and the focus are children of the
render-lines pane.

As a work around, move the target towards the root - as long as geo is
the same - so that a match is more likely.

Signed-off-by: NeilBrown <neil@brown.name>