search: more fixes to highlight when search fails
When search fails, we should still highlight other matches before the
cursor.
That doesn't always work because we don't always get a render-reposition
message to tell us what is visible.

So don't discard hi->start and hi->end so quickly, and when a search
fails, queue a delayed reposition notification, which will make use of
that ->start and ->end.

Signed-off-by: NeilBrown <neil@brown.name>