calculator pane
Add a full-precision calculator use some parsing code I wrote for
another project together with the gmp library, and utilize it
in a 'calculator' window which evaluates expressions and stores
results for use in later calculations.

In emac A-# replaces an expression under cursor with the evaluation,
in hex if the expr starts with '#'.

Interactive command "calc" enables calc mode for the document.

Signed-off-by: NeilBrown <neil@brown.name>