Introduce DEF_CB() similar to DEF_CMD()
There are two sorts of commands.
- those with a home pane that get state from ci->home->data
- those with no home that use container_of(ci->comm, ..., c)

This difference is not type-safe.  I'm not sure how to make it so, but
acknowledging it is good start.

In general, call() and pane_call() should result in a CMD being called;
comm_call() should result in a CB being called.
I'm not quite ready to enforce that though.

Signed-off-by: NeilBrown <neil@brown.name>