worddiff: start conversion to a command-object interface
The interface to diff/wiggle is already complex, and doesn't support
highlighting across multiple documents.

So start conversion to a command-object interface where a "wiggle"
command can be created and given the various pane and marks, then
asked to do its thing.

Only 'diff' handled so far, 'wiggle' is next.

Signed-off-by: NeilBrown <neil@brown.name>