diff: try harder to find correct line in target file.
If the target file has changed, the target line might have moved.
So when jumping from  line in a diff to the line in the original,
extract the 'after' part of the diff, and search for it near the
target location in the target file.

This involves adding a new "text-equals" command to lib-search,
and adding content to the "find" commands in lib-worddif.

Signed-off-by: NeilBrown <neil@brown.name>