lib-diff: simplify parsing of the diff output.
Use text-search to find the various markers, and extra subfields.
Use wiggle to parse the patch hunk and extract text.

New command "extract" is added to wiggle.

New interpolations \:N:M added to rexel to get a specific
instance of a group (not use yet, but needed soon).

Fix some bugs.

Refresh the diff test as this fixes a minor bug.

Signed-off-by: NeilBrown <neil@brown.name>