render-lines: use dedicated pane for off-screen cursor
When the cursor is not visible, we place it at the bottom-right, so it
can be found by the eye.

Currently this is drawn on the pane background, we never gets erased.

Instead, create a small pane, place it bottom-right, and draw the cursor
there.

This may overlap the last line on the pane, so we need to be able to
order the cursor pane on top of the normal content panes.

Signed-off-by: NeilBrown <neil@brown.name>