renderline: change rules for cursor placement on multi-glyph chars.
For TAB we want to leave the cursor on the first space drawn.
For a linewrap, we want to leave it on the last char (the char that
caused the wrap).
The current rule for this is a bit complex, and doen't work when a tab
is wrapped.
So simplify and base it on "in_tab".
This fixes a bug (for wrapped tabs) so a test needs to be refreshed.

Signed-off-by: NeilBrown <neil@brown.name>