Revert "notmuch: doc:step returns a char"
This reverts commit 8230fddc172f0c6863f5771d614c727369d31153.

This commit was wrong.  The value returned from doc:step
is being used as a return-value for doc:step, so we don't
want any transformation.

The one place where it is used locally, it is compared with edlib.WEOF
which is correct for an untransformed value.  The transformed value
would be None.

Signed-off-by: NeilBrown <neil@brown.name>