notmuch: always set tag changes to list_pane
I was optionally sending tag changes to the query pane if it existed.
I think the idea was that it might want to intercept and update cached
info, but it doesn't.
If it ever does, notifications would be a better approach, and we
already have Notify:Tag

So always sent to list-pane.

Signed-off-by: NeilBrown <neil@brown.name>