enhance get_str_callback
I'm going to use doc:content for all doc:get-str, so I want to be sure
it can be efficient.
So it can now pass expected size in ->x to minimize reallocation.
It can also pass multiple chars in ->str (->num2 bytes), but there is
no guarantee they will be used.

Signed-off-by: NeilBrown <neil@brown.name>