notmuch: fix mark-up when a threadid moves
When a threadid moves we want marks on that thread to be repositioned.
They are moved to before self.pos
If they are already immediately before self.pos, this causes an infinite
loop so we need to check for that case.

The old code which checked only covers one narrow case.  Remove that and
add more general code

Signed-off-by: NeilBrown <neil@brown.name>