rexel: handle ignore-case sections properly
As well as keeping the IgnoreCase state flag, we need to encode an
action to start ignoring case, and then an action to revert that
at the end of the section.  Add a test case for this.

Also fix up doco for the flags, and fix the implementation of capturing.
The 'n' flag applies to nested groups, while '?:' only applies to this
group.
However ?: does not stop capturing at the top level - nothing does.

Signed-off-by: NeilBrown <neil@brown.name>