doc-text: fix replace when endpoints are identical.
If the two endpoints text_replace is given, or finds, are exact the same
mark, then don't move it (as 'early') before other marks.  It is
important that the insert happens *after* those marks, and the insert
will in that case happen where the mark is moved to.

So in that case, make a duplicate for 'early'.

Signed-off-by: NeilBrown <neil@brown.name>