emacs-spell: make it possible to accept a spelling suggestion.
Record spelling suggestions and the location of the error and if alt-;
is used again at the same location, try a suggestion.

emacs-mode cannot hold marks, so it stores the current content (either
original error or accepted suggestion) both in the point and the pane.
If the text before point matches the content cached as an attribute in
point, assume that is the word to receive the next suggestion.

Signed-off-by: NeilBrown <neil@brown.name>