render-lines: improve setting of ->repositioned
This arose because moving up multiple lines did NOT trigger a
render:reposition message.  This was because a new vmark was created,
but didn't cause reposition to be set.

So fix that, and other improvements to ensure it is set properly.

Try to centralize some of the setting, but there is one place that
shouldn't be there, but removing it breaks search.  Need to research that.

Signed-off-by: NeilBrown <neil@brown.name>