spell: split out range handling code
The idea of a series of marks which divide the document in to ranges,
which can be added, deleted, or discovered, may be generally useful.

So split it out into add_range, remove_range and choose_range, which are
more general than the code I had.  They can be moved elsewhere when an
actual second use is found.

Signed-off-by: NeilBrown <neil@brown.name>