pane: suppress recursive notifications silently.
Recursive notifications can happen in sensible code.
At least: when a notification like doc:replaced is use for two things,
list change-text or change-attributes, and when the handler responding
to one thing causes another thing to be notifified, then that is a
recursive notification that can be suppressed, but shouldn't cause the
whole notification to fail.

So be more generous and don't fail if any recursion is detected.

Signed-off-by: NeilBrown <neil@brown.name>