discard "old style" scaling: x:XX,y:YY
My original design for scaling assumed x and y 'points' were the same
size.  An "M" would then be 10 points wide, by maybe not 10 points high.

This is't really easy to work with, as you cannot set a scale to
guarantee a minimum number of lines of the display.
So the new syle use "XXxYY" where units for XX are 1/10 the width of
default "M" and units for YY are 1/10 the default line height.
So "scale:800x240" will allow 80 x 24 Ms on the pane.

We don't really need "old style" any more, so remove it.

Signed-off-by: NeilBrown <neil@brown.name>