email-compose: enable posting of the message
The message is formated and passed to sendmail, or other program as
specified by the caller.
Notmuch configures "notmuch insert" to be used.  The plan it to have a
job that looks for tag:outbox messages and submits them, removing
tag:outbox on success.

Empty headers are removed, and Date: and Message-id are added if needed.

Signed-off-by: NeilBrown <neil@brown.name>