doc-multipart: fix incorrect failure of doc:step-part
doc:step-part fails if it was asked to move to a different part, but
ended up on the part it started on.
Unfortuantely it also fail if it was asked to move to the start of the
current part, but ended up on the part it started on!!

So fix the error case.

Signed-off-by: NeilBrown <neil@brown.name>