notmuch: remove a level of indirection in make_search()
make_search() currently prefixes the search name with 'saved:', then
expands it - which will strip the 'saved' and do a lookup.

This requires that each search name matches the given regexp, and has
already been a minor problem for the "tag:" non-saved search.
I'm going to introduce another non-saved search, so remove the
indirection and do the lookup directly in make_search()

Signed-off-by: NeilBrown <neil@brown.name>