Change doc:content to pass mark *after* char to callback.
The callback now received:
 - the next char/byte
 - the mark *after* that char/byte
 - optionally some more bytes/chars in ->str/->num2

This makes it easier for filters which make multi-chars into a single
char to report the correct marks.

For base64, neither just-before or just-after is particularly easy as a
char is made of bit of two adjacent b64 chars in a sliding window.  It
was wrong before.  Now it is right.

Signed-off-by: NeilBrown <neil@brown.name>