Introduce doc:char
doc:char is the new doc:step.  It can take a repeat count, and uses
negative for backwards and positive for forwards.
It can also stop at a given more so that distances can be counted.

For now, everything that implemented doc:step gets a uniform doc:char
implementation which wraps doc:step.  Gradually that will change.

Signed-off-by: NeilBrown <neil@brown.name>