notmuch: use pane attributes to store tid/mid of a message view
Currently the tid/mid of a message being viewed are stored in python
object attributes.  This means that non-python code in some other pane
would not be able to access them.
I prefer that between panes, only edlib interfaces are used.
So instead store the thread-id/message-id in pane attributes.

Signed-off-by: NeilBrown <neil@brown.name>