attr: fix incorrect allocation
One case for allocating more attr space was broken.
If we decided to split an attrset and store the new attr in the first
half, we only confirmed that the total size would be less than the MAX,
not that the required size had actually been allocated.
So loop around and re-test.

Also handle the case of len exceeding MAX better.  When offset is zero
and resizing the current attrset won't be allowed, allocate a new
attrset just for this attr.

Signed-off-by: NeilBrown <neil@brown.name>