make: continue following when point moved to EOF
The "may_follow" variable is needed only to suppress a single
follow-the-EOF even.  Once one has been skipped others will
automatically be ignored because point is not at EOF.  If point is ever
moved to EOF we want it to be followed again.
So clear may_follow when we suppress a follow operation.

Also skip the doc:replaced handling if num2 is 1, meaning that only
attributes were changed.

Signed-off-by: NeilBrown <neil@brown.name>