Introduce mark_valid()
When a mark is freed (mark_free()) it is marked invalid and only freed
at the next idle time.  This is because some code might still have a ref
on it.
To make full use of these, we need to test if a mark is still valid
before using it.
So introduce mark_valid() and use it in core-mark and lang-python.

Also other misc cleanups.

Signed-off-by: NeilBrown <neil@brown.name>