Automatically close all vmarks when pane is closed.
When either the document or the owner of a vmark is closed, we now
automatically free all the vmarks.  If the vmark as ->mdata then
"Close:mark" is called on the owner to free up the ->mdata.
If that doesn't do the right thing, a warning is generated and the
data is ignored (probable memory leak).

points were already freed when the doc was closed.

ungrouped marks with ->mdata cannot be handled transparently so again a
warning is generated.

Signed-off-by: NeilBrown <neil@brown.name>