make: abort timer when triggering a reparse.
When an non-neg arg is given to 'match-next' it purges all parse info
to restart the parse.  But if the timer is still ticking, it doesn't
immediately reparse.  This is unlikely in normal use, but happens
reliably in the test harness.

It makes sense to cancel the timer in this circumstance, so do that.

Signed-off-by: NeilBrown <neil@brown.name>