notmuch master_view: don't send provided mark to other panes
When master_view redirects a comment to one of the sub-panes (line,
query, message), it mustn't send the mark it received as it could be for
a different document.  It can only use that mark which interacting with
focus.

So when explicitly using a chosen subpane, get the doc:point for that
pane.

Signed-off-by: NeilBrown <neil@brown.name>