lib-input: support fallback selection owner.
A selection owner can now request (with ci->num == 1) to be a fallback.
If the selection owner is ever lost, it reverts to the fallback instead
of to NULL.

This allows x11selection-xcb to be sure that every display always has a
selection owner, so that any attempt to commit the selection will fetch
PRIMARY if nothing else is appropriate.

Signed-off-by: NeilBrown <neil@brown.name>