lib-input: allow a selection:commit to not set ->sel_committed
Normally when we when the selection is committed, it stays committed and
we don't commit it again until it is re-asserted.
However the X11 selection protocol doesn't signal a re-assertion (unless
we take ownership, which would be inappropriate).
So when x11selection handles a commit, it needs to be able to ask for
->sel_committed to remain unset so it can get called again.

This doesn't result in the same selectin being added to the copy history
multiple times, as we don't actually add it from x11selection-xcb.

Signed-off-by: NeilBrown <neil@brown.name>