notmuch: decouple count updates and query updates.
Currently we update all the counts, then all the queries.
There is no need for the strict sequencing, particularly as some counts
are already parallel to handle slow searches.

So have a separate sequence for query update and whenever a count shows
the a query has changed, schedule an update for that query immediately.

If it is visible, do the query.  If not, mark the query for update when
next visited, even when it is under 60 seconds since last time.

Signed-off-by: NeilBrown <neil@brown.name>