display: record when pane is a uniform background colour.
When a pane is a uniform background colour, and is used to clear another
pane, we can just copy the colour, not the content.
This is probably more efficient, and avoids problems with partial
over-laps of panes.

Also: only use parent information for 'Draw:clear' when attrs is NULL.
If it is non-null but contains no bg, use white.

Signed-off-by: NeilBrown <neil@brown.name>