render-format: always used '0' for start-of-line index.
If the first field of the format sets attributes, then the normalized
start of the line won't be zero.
But we really need ref.i to be zero at start of line, as sometimes
the mark is moved outside our control - always to ref.i = 0.

So treat '0' and 'start of line' as special, and don't normalize the
same way.

Signed-off-by: NeilBrown <neil@brown.name>