display-x11: only redraw content that has changed.
Use cairo_region to track where expose-events have told us to redraw,
and to track which panes we have changed ourselves.

Then only redraw those.

As part of this, reverse the sort order of 'panes' so that we handle the
high-z panes first, and subtract them from the remaining region.  This
ensure we never draw the same pixel twice for a single refresh.

Signed-off-by: NeilBrown <neil@brown.name>