test: change the handling of EDLIB_REPLAY_DELAY
Previously EDLIB_REPLAY_DELAY would add a short pause so that someone
running "tests view ..." should see what has happening.
It did that with usleep()

Now, instead, it uses event:timer to delay the next-event call a little.
This means that updates can continue and delayed updates - like
spell-check finding more errors - will get seen.

When the is a display update - extent the time a little more to be sure
we get the next one.

Now we enable the REPLAY_DELAY when running 'tests refresh...'

ALSO: fflush writes to the RECORDING so I can monitor progress more
easily
ALSO: change shell more to slurp in the whole result after a short
pause, so that the result is quite stable.

Signed-off-by: NeilBrown <neil@brown.name>