Forbid nested notification again.
Nested notifications really are bad.  The original notification will not
get everywhere.
Nested notifications aren't the solution for doc:replaced having two
uses.
Just not doing that is the solution.
So now we have doc:replace-attr which needs to be requested separately
if wanted.

This particularly affects auto-spell, which now has a lot of noised
removed from the test script.

Signed-off-by: NeilBrown <neil@brown.name>