render-complete: fix render_complete_line
render_complete_line assumes it is rendering the whole line.  It uses
this assumption to determine where the content to highlight is.

But it might only be rendering to a mark, or to an offset.  The match
content might not be present, or might be only partially present.

So always render the whole line first.  Then, if necessary, adjust to
the the requested details.

Signed-off-by: NeilBrown <neil@brown.name>