notmuch: update queries when not displayed.
I don't like the small delay when visiting a query that has new
messages.
But I don't want to pay the price of always updating in the background
even when I'm not looking at a query.
So track updates for the first 5 minutes, then assume the query
isn't interesting and stop updating.

Signed-off-by: NeilBrown <neil@brown.name>