Add global-config-dir for make commands
dir-specific config can now be attached to the root and accessed with
global-config-dir.
The value and dir in which it was found is returned.

Make uses this to find a command, and to find the directory to run it
in.
A '%%' in the command is replaced with the part of the directory that
was stripped off.

Signed-off-by: NeilBrown <neil@brown.name>