notmuch: keep point nearby when returning from whole-thread
If, while viewing a whole thread, point is moved to EOF, then when we
switch back we don't want it to stay at EOF, we want it to be at
thread_end, or similar.
So we need to clip anything after thread_end upwards.

Clip doesn't move marks *at* the end point, so we need to make our eof
look just a little bit different temporarily.

Signed-off-by: NeilBrown <neil@brown.name>