Add ref-counting for panes
New pane_get() and pane_put() make it possible to ensure a pane is not
freed while a reference is held.  This is used by python to stop bad
code from crashing when holding an old ref to a pane.

The pane can be closed and the ->data can be freed, but the pane object
itself will remain so that it can be safely tested to see if it is valid.

Signed-off-by: NeilBrown <neil@brown.name>