notmuch: messages in visible thread should not disappear.
While a thread is visible, background refresh should not cause messages
to disappear - e.g. if I archive one and then a refresh happens.
They should only disappear when I re-visit the thread, or when I
explicitly request a refresh with '='.

So keep track of that explicit request, and when not present but the
query is open, preserve all 'True' matched values.

Signed-off-by: NeilBrown <neil@brown.name>