render-lines: don't change y_pre or y_post after found_start/end set
To be able to achieve precise positioning on the display (e.g.  with
vline) we need to stop changing y_pre or y_post once found_start or
found_end are set - respectively.

So when we hit EOF and set a 10% y_post, we should NOT set found_end
yet.

This shows an off-by-one error in vline calculation which was previously
hidden as consume_space() would always move us to the start of that line.

Signed-off-by: NeilBrown <neil@brown.name>