Use XDG standard to find config (and data) files.
Provide an interface to search for config and data files in appropriate
XDG locations.  Enhance this by looking in the directory that contains
the main shared library - after HOME but before system locations.
Use this for config files.  This removes the need to use ARGV[0]
to find the initial ini file.

Signed-off-by: NeilBrown <neil@brown.name>