Support inline pane data.
If a NULL data pointer is passed to __pane_register(), then a larger
memory allocation will be made to hold both the pane and the data.

If a type is defined with PANE_DATA_TYPE then that will be the size
passed to __pane_register by pane_register().

Similarly DOC_DATA_TYPE can give a type for a document.  The "struct
doc" must appear first in this type.

If PANE_DATA_TYPE is defined, then "core-pane.h" must be included after
the type has been declared, which will usually be after "core.h" is included.

Signed-off-by: NeilBrown <neil@brown.name>