Add lib-askpass
AskPass will as for a password and not display what is entered.
A new script "el-askpass" is provided that ssh can use to ask for a
password through edlib.
lib-shellcmd sets the environment so this script is used.

Net result is that if a command run from edlib needs ssh to ask for a
password, it can.

Signed-off-by: NeilBrown <neil@brown.name>