Change how Move-CursorXY is passed "no change" co-ordinates.
Currently a negative co-ordinate means "use the ->cx/cy value of the target
pane".  This allows changing position in one dimension only.

However it prevents receiving a negative co-ordinate, which might happen
in the mount event happens out-side the pane which has grabbed the
mouse.

So instead of a negative number, use MAX_INT to say "don't change".

Signed-off-by: NeilBrown <neil@brown.name>