massive lib-renderline rewrite
This now passes my tests - after adjusting some tests.

lib-renderline is totally different.  Parsing and processing are now
distinct steps.  Hopefully this will make it easier to work with in
future, though it is still very rough in places, particularly with
handling for \t tabs.

I need to go through it several times and clean stuff up, but at least I
now have a credible base to work on.

Signed-off-by: NeilBrown <neil@brown.name>