tests: Allow extra data files to be used in tests.
Only files in tests.d that start with a digit are now considered to be
tests.
Other files are copied into DOC/extra in the test area for tests to
examine.
I put it DOC so that most tests don't see the extra entry in a directory
and so don't need to be changed.  Only 02-presenter looks in DOC

Signed-off-by: NeilBrown <neil@brown.name>