Allow command name to be used to choose some data files.
The icon file, Welcome text, and initial config file are now chosen
based on the base name of the command being run.
So if "edlib" is copied to "elma", then "elma.ini" is used for config,
and "elma-icon.png" for an icon.
If the expected file isn't found, a file named for "edlib" is tried
instead.

This should make it easy to have have various commands providing
different interfaces.
 elma - EMACS mode
 elvi - VI mode
 elnm - notmuch email reader
 elpnt - presentation mode
 eled - editor with A binding.
maybe.

Signed-off-by: NeilBrown <neil@brown.name>