ncurses: fix setting of default attrs on Draw:clear
If no attrs are given for Draw:clear, we use background from parent.

This wasn't implemented properly.  If the pane receiving the text was
the same as the pane that was cleared, the text got the background of a
different pane.  That should only happen on clear, not when drawing text.

Signed-off-by: NeilBrown <neil@brown.name>