menu: support detaching the doc from a menu.
It might be use to create a menu doc once, and re-use it.

For this to work we need to be able to create a menu with an existing
doc.
A new mode for attach-menu allows this.

It will also be useful to be able to edit the detached document so add
global commands menu:add and menu:clear which do this.

Signed-off-by: NeilBrown <neil@brown.name>