Introduce closed_ok flag on commands.
If closed_ok is set, then it is safe to call the command even if ->home
is DAMAGED_CLOSED.  This will eventually replace checks on "Close".
For now it allows doc_pass_on to pass commands on to the doc even
when the display stack is shutting down.  notmuch needs this to ask
the query doc to refresh when a view is closed.

Signed-off-by: NeilBrown <neil@brown.name>