Remove meaning for str2 arg to Draw:image and Draw:image-size
The 'str2' arg is passed to the command when the image name start
"comm:".
However this was never use and it is not clear that the caller could
ever have a useful string to pass.  All details of how to find the
images can be encoded in the command name.

So drop all use for str2 in Draw:image and Draw:image-size handlers.

Signed-off-by: NeilBrown <neil@brown.name>