Split display creation into two separate steps.
This effectively reverts an earlier patch.
That patch caused x11 selection to stop working because
the relevant pane was attached before DISPLAY had been set.
So now we create the display first, then have the opportunity to set
attributes, then call window-active-display which adds all the extra
panes and attaches a document.

Signed-off-by: NeilBrown <neil@brown.name>